OpenMP: Remove KOKKOS_ENABLE_OPENMP macro.
