+++ core/src/OpenMP/Kokkos_OpenMP.hpp
@@ -23,7 +23,6 @@ static_assert(false,
 #define KOKKOS_OPENMP_HPP
 
 #include <Kokkos_Macros.hpp>
-#if defined(KOKKOS_ENABLE_OPENMP)
 
 #include <Kokkos_Core_fwd.hpp>
 
@@ -36,7 +35,6 @@ static_assert(false,
 #endif
 
 #include <Kokkos_ScratchSpace.hpp>
-#include <Kokkos_Parallel.hpp>
 #include <Kokkos_TaskScheduler.hpp>
 #include <Kokkos_Layout.hpp>
 #include <impl/Kokkos_HostSharedPtr.hpp>
@@ -234,14 +232,9 @@ struct MemorySpaceAccess<Kokkos::OpenMP::memory_space,
 
 #include <OpenMP/Kokkos_OpenMP_Instance.hpp>
 #include <OpenMP/Kokkos_OpenMP_Team.hpp>
-#include <OpenMP/Kokkos_OpenMP_Parallel.hpp>
-#include <OpenMP/Kokkos_OpenMP_Parallel_For.hpp>
-#include <OpenMP/Kokkos_OpenMP_Parallel_Reduce.hpp>
-#include <OpenMP/Kokkos_OpenMP_Parallel_Scan.hpp>
 #include <OpenMP/Kokkos_OpenMP_Task.hpp>
 
 #include <KokkosExp_MDRangePolicy.hpp>
 /*--------------------------------------------------------------------------*/
 
-#endif /* #if defined( KOKKOS_ENABLE_OPENMP ) && defined( _OPENMP ) */
 #endif /* #ifndef KOKKOS_OPENMP_HPP */
