+++ core/src/OpenMP/Kokkos_OpenMP_Parallel_Reduce.hpp
@@ -18,9 +18,9 @@
 #define KOKKOS_OPENMP_PARALLEL_REDUCE_HPP
 
 #include <Kokkos_Macros.hpp>
-#if defined(KOKKOS_ENABLE_OPENMP)
-
-#include <OpenMP/Kokkos_OpenMP_Parallel.hpp>
+#include <omp.h>
+#include <OpenMP/Kokkos_OpenMP_Instance.hpp>
+#include <KokkosExp_MDRangePolicy.hpp>
 
 namespace Kokkos {
 namespace Impl {
@@ -565,5 +565,4 @@ class ParallelReduce<CombinedFunctorReducerType,
 }  // namespace Impl
 }  // namespace Kokkos
 
-#endif /* KOKKOS_ENABLE_OPENMP */
 #endif /* KOKKOS_OPENMP_PARALLEL_REDUCE__HPP */
