+++ core/src/OpenMP/Kokkos_OpenMP_Parallel_Scan.hpp
@@ -14,13 +14,12 @@
 //
 //@HEADER
 
-#ifndef KOKKOS_OPENMP_PARALLEL_REDUCE_SCAN_HPP
-#define KOKKOS_OPENMP_PARALLEL_REDUCE_SCAN_HPP
+#ifndef KOKKOS_OPENMP_PARALLEL_SCAN_HPP
+#define KOKKOS_OPENMP_PARALLEL_SCAN_HPP
 
 #include <Kokkos_Macros.hpp>
-#if defined(KOKKOS_ENABLE_OPENMP)
-
-#include <OpenMP/Kokkos_OpenMP_Parallel.hpp>
+#include <omp.h>
+#include <OpenMP/Kokkos_OpenMP_Instance.hpp>
 
 //----------------------------------------------------------------------------
 //----------------------------------------------------------------------------
@@ -311,5 +310,4 @@ class ParallelScanWithTotal<FunctorType, Kokkos::RangePolicy<Traits...>,
 }  // namespace Impl
 }  // namespace Kokkos
 
-#endif /* KOKKOS_ENABLE_OPENMP */
 #endif /* KOKKOS_OPENMP_PARALLEL_REDUCE_SCAN_HPP */
